fix tests

+++ scripts/ci-test.sh
@@ -2,6 +2,11 @@
 
 set -e
 
+if [ -f peterbecom/settings/local.py ]; then
+  echo "The file peterbecom/settings/local.py already exists. Refusing to override."
+  exit 1
+fi
+
 echo "Making settings/local.py"
 cat > peterbecom/settings/local.py <<SETTINGS
 from . import base
